+++ e2e-spa/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @scale-codec/e2e-spa
 
+## 0.0.4
+
+### Patch Changes
+
+-   Updated dependencies [undefined]
+    -   @scale-codec/definition-runtime@0.6.0
+    -   @scale-codec/definition-compiler@0.4.0
+
 ## 0.0.3
 
 ### Patch Changes